requirements.txt
torch
torchvision
pillow
matplotlib
opencv-python